requirements.txt
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pydantic==2.5.3
pyyaml==6.0.1
flask==3.0.0
